Compute ManifestUpdateOptions.InformationOnly.LayerDiffIDs
This is pretty ugly (decompressing the whole tarballs only to compute a
digest), but necessary to convert from schema1 to schema2, where this
information must be embedded into a config.json.

Due to the performance impact (it completely defeats PutBlob detecting
that a layer is already present on the destination and not reading it),
we do this only if the horribly ugly Image.UpdatedImageNeedsLayerDiffIDs
says so.

Right now, the code exists but is essentially deactivated because
UpdatedImageNeedsLayerDiffIDs just returns false in both
genericManfiest implementations.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>