API transition: Drop XNamed.XName
Instead call distreference.FamiliarName() in SOME uses.

In signature/policy_reference_match.go, where we care about equality but
not exactly about the kind of normalization, call XNamed.Name()
instead.  That compares the fully-explicit instaed of the
fully-minimized name forms.  If both canonicalizations are consistent,
this should not matter—and if they weren’t, using the fully explicit
form should be safer.  (Also, .Name() is likely to be a bit faster, but
that really doesn’t matter all that much.)

Signed-off-by: Miloslav Trmač <mitr@redhat.com>