Use paranoidUnmarshalJSONObjectExactFields in policy.json parsing
In most of the cases, the original code did not explicitly require that
a (string) field is present, but empty strings were later rejected
(because the "type" field was required to be equal to a constant, or
because parsing a docker/reference would fail).  Either way, there were
already tests verifying that missing fields are rejected; now the code
is just a bit more explicit about it.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>