storage images: use verbose names on images
When writing an image to the storage transport, add an image name based
on the tag, if one was given.  This should save most consumers of the
library from having to mess around with names after calling
copy.Image().

When reading or deleting an image, make sure that if the reference we're
using to find it includes a name, that the image has the name, for the
sake of cases where the reference was initialized with both a name and
an ID.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>